Add a process of resetting i/o buffers when opening a serial port
There was an issue where if data reached the serial buffer before the
server was started, the server would read that data upon startup. This
problem could cause the server to receive invalid data if the client
sent data first and timed out, leading to potential improper behavior.

To avoid this, a process to clear the buffer was added when the serial
port is opened.